Fix startup process twice #935
The initialization process that create cheshire_cat_api is called twice

Uvicorn.run with the reload options true, run main.py again

Moved the the initialization tasks in new startup.py file
